google-speech-examples/requirements.txt
beautifulsoup4==4.9.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-speech==1.3.2
googleapis-common-protos==1.51.0
grpcio==1.28.1
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
